WIP-23-09-2023-21-12-52

+++ supabase/migrations/20230916151331_get_similar_houses.sql
@@ -8,7 +8,7 @@ RETURNS TABLE (
   house_price numeric,
   total_required numeric,
   personal_invested_amount numeric,
-  amount_reached int4
+  amount_reached numeric
   ) AS $$
 BEGIN
     RETURN QUERY
